.Net: Splitting semantic and native function internal implementations (#2349)
### Description

Splitting semantic and native functions into `internal` implementations
of ISKFunction. This helps to isolate the (largely independent) code
paths and bring clarity to the internal implementation.

This change does NOT introduce breaking changes; all public properties
and methods are preserved.

Note: In a future change, the interface and properties will be further
edited, and that will be a breaking change. This PR is a first pass, and
aims to clean things up as much as possible with zero impact to the
public API surface.

Steps ahead:
- Removing methods from the public ISKFunction interface if they only
apply to one type
 - Adding helpers to identify and cast to the correct type
- Splitting semantic functions into its own package -- the
SemanticKernel.Core/Abstractions packages will have no direct knowledge
of prompt templates or semantic functions as it will be exposed as a
plugin package (that functionality will remain available by default for
anyone using the `Microsoft.SemanticKernel` metapackage).

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [X] The code builds clean without any errors or warnings
- [X] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [X] All unit tests pass, and I have added new tests where possible
- [X] I didn't break anyone :smile:
